fix + refactor: Fix tests and refactor
