+++ apiary.apib
@@ -363,9 +363,9 @@ Update informations of the current Customer.
 
 + Response 204
 
-### Delete Profile of the current Customer [PATCH]
+### Delete Account of the current Customer [DELETE]
 
-Delete informations of the current Customer.
+Delete the current Customer account.
 
 **Endpoint information**
 
@@ -377,6 +377,7 @@ Delete informations of the current Customer.
     + Headers
 
             Authorization: Bearer <JWT>
+
 + Response 204
 
 ## Data Structures
