+++ src/application/commands/customer/delete-customer-by-uuid.command.ts
@@ -1,7 +1,7 @@
 import {ICommand} from '@nestjs/cqrs';
 import {Customer} from '@letseat/domains/customer/customer.entity';
 
-export class DeleteCustomerByUuidQuery extends Customer implements ICommand {
+export class DeleteCustomerByUuidCommand extends Customer implements ICommand {
 	readonly uuid: string;
 
 	constructor(uuid: string) {
