+++ src/application/commands/customer/index.ts
@@ -1,2 +1,3 @@
 export {CreateCustomerCommand} from './create-customer.command';
 export {UpdateCustomerCommand} from './update-customer.command';
+export {DeleteCustomerByUuidCommand} from './delete-customer-by-uuid.command';
