+++ src/application/queries/customer/handlers/index.ts
@@ -2,20 +2,17 @@ import {GetCustomerByEmailHandler} from './get-customer-by-email.handler';
 import {GetCustomerPasswordHandler} from './get-customer-password.handler';
 import {GetCustomerByUuidHandler} from './get-customer-by-uuid.handler';
 import {GetCustomersHandler} from './get-customers.handler';
-import {DeleteCustomerByUuidHandler} from '@letseat/application/queries/customer/handlers/delete-customer-by-uuid.handler';
 
 export const CustomerQueryHandlers = [
 	GetCustomerByEmailHandler,
 	GetCustomerPasswordHandler,
 	GetCustomerByUuidHandler,
-	GetCustomersHandler,
-	DeleteCustomerByUuidHandler
+	GetCustomersHandler
 ];
 
 export {
 	GetCustomerByEmailHandler,
 	GetCustomerPasswordHandler,
 	GetCustomerByUuidHandler,
-	GetCustomersHandler,
-	DeleteCustomerByUuidHandler
+	GetCustomersHandler
 };
