+++ src/application/queries/customer/index.ts
@@ -1,4 +1,3 @@
 export {GetCustomerByEmailQuery} from './get-customer-by-email.query';
 export {GetCustomerPasswordQuery} from './get-customer-password.query';
 export {GetCustomerByUuidQuery} from './get-customer-by-uuid.query';
-export {DeleteCustomerByUuidQuery} from './delete-customer-by-uuid.query';
